# src/main/webui/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    ChartNode:
      type: object
      required:
      - index
      - value
      properties:
        index:
          type: string
        value:
          type: string
    ExternalProductQueryMessage:
      type: object
      properties:
        keyword:
          type: string
        sources:
          type: array
          items:
            type: string
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
    PeriodLength:
      type: string
      enum:
      - WEEK
      - MONTH
      - QUARTER
      - YEAR
    ProductInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          minLength: 1
        imgUrl:
          type: string
        desc:
          type: string
        gtin:
          type: string
          maxLength: 14
          minLength: 8
          pattern: "^([0-9]{8}|[0-9]{12,14})$"
        unit:
          type: string
          default: PER_UNIT
          $ref: "#/components/schemas/Unit"
        latestQuotes:
          type: array
          items:
            $ref: "#/components/schemas/SimpleQuoteDTO"
          readOnly: true
        stats:
          type: object
          additionalProperties:
            type: number
          readOnly: true
        quantityPerItem:
          type: number
        itemPerBundle:
          type: integer
          format: int32
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
      required:
      - name
      - unit
    QuoteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        productInfo:
          $ref: "#/components/schemas/ProductInfo"
        storeInfo:
          type: object
          anyOf:
          - $ref: "#/components/schemas/StoreInfo"
          - $ref: "#/components/schemas/WithId"
        quoteDate:
          type: string
          $ref: "#/components/schemas/LocalDate"
        price:
          type: number
        discountType:
          $ref: "#/components/schemas/Type"
      required:
      - storeInfo
      - quoteDate
      - price
    QuoteResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/QuoteDTO"
    SimpleQuoteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        storeInfo:
          $ref: "#/components/schemas/StoreInfo"
        quoteDate:
          type: string
          $ref: "#/components/schemas/LocalDate"
        price:
          type: number
        discountType:
          $ref: "#/components/schemas/Type"
      required:
      - quoteDate
      - price
    StoreInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        address:
          type: string
        url:
          type: string
      required:
      - name
    TimeRange:
      type: string
      enum:
      - ONE_MONTH
      - THREE_MONTH
      - SIX_MONTH
      - TWELVE_MONTH
    Type:
      type: string
      enum:
      - PERCENTAGE
      - FIXED_AMOUNT
      - BUNDLE
      - NEAR_EXPIRY
      - OTHER
    Unit:
      type: string
      enum:
      - EA
      - KG
      - ML
      - G
      - L
      - M
    WithId:
      required:
      - id
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
      type: object
paths:
  /api/common/latestDeals:
    get:
      parameters:
      - name: s
        in: query
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResult"
      summary: Get Recent Discount
      tags:
      - Pricekeep Resource
  /api/product/alerts:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteDTO"
      summary: Get Today Alert Trigger
      tags:
      - Product Resource
  /api/product/all:
    get:
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductInfo"
      summary: Get All Product
      tags:
      - Product Resource
  /api/product/extSources:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      summary: Get Available Ext Sources
      tags:
      - Product Resource
  /api/product/gtin/{gtin}:
    get:
      parameters:
      - name: gtin
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductInfo"
      summary: Get Product By Gtin
      tags:
      - Product Resource
  /api/product/new:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInfo"
        required: true
      responses:
        "201":
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductInfo"
        "400":
          description: Data conflict
          content:
            application/json: {}
      summary: Create Product
      tags:
      - Product Resource
  /api/product/search:
    get:
      parameters:
      - name: q
        in: query
        schema:
          type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductInfo"
      summary: Search Product
      tags:
      - Product Resource
  /api/product/searchExt:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalProductQueryMessage"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteDTO"
        "400":
          description: Bad Request
      summary: Find External Product
      tags:
      - Product Resource
  /api/product/suggest:
    get:
      parameters:
      - name: q
        in: query
        schema:
          type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      summary: Suggested Words
      tags:
      - Product Resource
  /api/product/{productId}:
    get:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductInfo"
      summary: Get Product By Id
      tags:
      - Product Resource
  /api/product/{productId}/chart:
    get:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: showDiscount
        in: query
        schema:
          type: boolean
      - name: tr
        in: query
        schema:
          $ref: "#/components/schemas/TimeRange"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartNode"
      summary: Get Chart Data
      tags:
      - Product Resource
  /api/product/{productId}/edit:
    post:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInfo"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Edit Product
      tags:
      - Product Resource
  /api/product/{productId}/quote:
    post:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Quote Price
      tags:
      - Product Resource
  /api/product/{productId}/quoteHist:
    get:
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: discount
        in: query
        schema:
          type: boolean
      - name: l
        in: query
        schema:
          $ref: "#/components/schemas/PeriodLength"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteDTO"
      summary: Get Price History
      tags:
      - Product Resource
info:
  title: pricekeep API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value