# src/main/webui/src/model/unit.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * pricekeep API
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 */

export type Unit = (typeof Unit)[keyof typeof Unit];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Unit = {
  EA: "EA",
  KG: "KG",
  ML: "ML",
  G: "G",
  L: "L",
  M: "M",
} as const;